default dummy() ISR : endless loop instead of just returning.
Interrupts on SuperH set the prio mask in SR to the current int being
serviced, so higher-priority ints can still be processed. So, multiple
dummy() calls are possible. An example call stack due
to a software bug (unwanted interrupts) could look like this :

3: dummy() called from NMI (prio=16)
2: dummy() called from high-priority vector (ex.: UBC, prio=15)
1: dummy() called from an unwanted, low-priority vector (ex.: prio=8)
0: user code

so the code will still fail and the ECU will restart.

A fancier version would instead write a global system_status variable with
the offending int, but that doesn't help if
- AUD unavailable
- kernel bug that prevents serial comms

which covers almost every possible scenario.