350nm: incomplete C implementation of Erase ucode.
First step is a rewrite of Renesas FDT kernel code in C, based on the
SHC flavor of assembly. Converting to GNU AS was attempted, but sucks
because of no automatic literal pool management. Also, gcc -Os is
competitive against the Renesas asm code.

One weakness of the C implem is timing; I have only average confidence
in the current waitn() implem.