Many fixes to functions.h and extra_functions.h
Changed implementation of imask_savedisable() and imask_restore():
-- uses less scratch registers
-- shorter
-- imask_restore() sometimes broke where the same reg was used as the
input as well as the tmp scratch. Not sure why the constraints allowed
gcc to do this (or maybe a bug) but this bypasses the problem
completely.

- pl_flash 180nm : use imask_savedisable + restore

- fixed "trapa" : can only take an immediate operand. Never caused a
  problem but was incorrect

- commented out unused / suspicious inline funcs

- rewrote functions as single-block `asm volatile (....)` statements to
  prevent gcc from reordering
(caused issues in GH #7 and in local builds)

For example, get_vbr() prior to mods, written as two separate
asm("....") lines :

.text:FFFF8C5A 02 22                       stc     vbr, r2
.text:FFFF8C5C 92 80                       mov.w   word_FFFF8D60, r2 ;
Clobbered VBR !!
.text:FFFF8C5E 2F 22                       mov.l   r2, @r15

gcc inserted an unrelated mov.w in the middle, for no reason.

These implementations should be more robust.